fgfgf

+++ Scripts/oraclecloud/jarbas.sh
@@ -1,7 +1,7 @@
 #######################################################
 ## jarbas Oracle Cloud ################################
 #! /bin/sh
-VERSION="v1.09"
+VERSION="v1.11"
 echo "========================================"
 case "$1" in
     ligar)
@@ -92,23 +92,27 @@ case "$1" in
               ./jarbas ligar
             ;;        
             instalarzip)
-              echo "===== Instala Foundry VTT ====="
-              ./jarbas desligar
-              echo "Voce deve ter feito o upload do arquivo zip do Foundry VTT (node) para a pasta principal atraves do Filezilla." 
-              echo "NOME DO ARQUIVO DEVE SER: fvtt.zip"
-              echo "Se nao fez, cancele esse processo com Control+C. Se teve problemas pode instalar via link com o ./jarbas instalar LINK conforme e ensinado no video: https://youtu.be/Y1Nw2B4NvKM?t=1081"
-              echo "Sua instalacao Foundry VTT esta sendo removida. Seus arquivos de usuario nao serao afetados."
-              cd ~
-              rm -rf foundry
-              mkdir -p foundry  
-              mkdir -p ~/.local/share/FoundryVTT  
-              mv foundry.zip foundry/
-              cd foundry/  
-              unzip foundry.zip
-              sudo chmod +x ~/foundry/resources/app/main.js    
-              rm foundry.zip
-              cd ~
-              ./jarbas ligar 
+              echo "===== Instala Foundry VTT ====="              
+              FILE=foundry.zip
+              if [ -f "$FILE" ]; then
+                  echo "$FILE localizado. Processo iniciado!"
+                  ./jarbas desligar
+                  echo "== Sua instalacao Foundry VTT esta sendo removida. Seus arquivos de usuario nao serao afetados. =="
+                  cd ~
+                  rm -rf foundry
+                  mkdir -p foundry  
+                  mkdir -p ~/.local/share/FoundryVTT  
+                  mv foundry.zip foundry/
+                  cd foundry/  
+                  unzip foundry.zip
+                  sudo chmod +x ~/foundry/resources/app/main.js
+                  rm foundry.zip
+                  cd ~
+                  ./jarbas ligar
+                else
+                  echo "Voce deve ter feito o upload do arquivo zip do Foundry VTT (node) para a pasta principal atraves do Filezilla."
+                  echo "NOME DO ARQUIVO DEVE SER: foundry.zip"
+              fi 
             ;;
             instalar)
               echo "===== Instala Foundry VTT ====="
@@ -132,6 +136,7 @@ case "$1" in
             echo "Exemplo de uso: ./jarbas admin removesenha"
             echo
             echo "instalar: instala uma nova versao do foundry vtt usando o link do site do foundry vtt. Tem que usar o link do node.js. Importante: a pasta da instalacao foundry atual sera apagada, isso nao afeta a pasta de dados do foundry."  
+            echo "instalarzip: instala o foundry a partir de um arquivo ZIP cujo nome dever ser foundry.zip e esse arquivo deve estar na pasta inicial no filezilla. Esse processo remove sua instalacao foundry sem afetar os arquivos de usuario."  
             echo "removesenha: remove a senha do foundry vtt"  
             echo "resetaconfig: coloca o arquivo de configuracao do foundry vtt em seu estado padrao."               
             echo
